feat: add toggleable dietary guide
